(Sands Fish) Provides fix for AuthenticationUtil where users ID's are not properly compared. As corrects issue with logged in user not having appropriate rights to be in this particular flowscript once logged in as. Eliminates need for user to do "double-logout".
git-svn-id: http://scm.dspace.org/svn/repo/branches/dspace-1_5_x@3006 9c30dcfa-912a-0410-8fc2-9e0234be79fd

+++ dspace-xmlui/dspace-xmlui-api/src/main/java/org/dspace/app/xmlui/utils/AuthenticationUtil.java
@@ -377,7 +377,7 @@ public class AuthenticationUtil
     	    Integer effectiveID = (Integer) session.getAttribute(EFFECTIVE_USER_ID); 
     	    Integer authenticatedID = (Integer) session.getAttribute(AUTHENTICATED_USER_ID); 
     	    
-    	    if (effectiveID != authenticatedID)
+    	    if (effectiveID.intValue() != authenticatedID.intValue())
     	    {
     	    	// The user has login in as another user, instead of logging them out, 
     	    	// revert back to their previous login name.
